proto/info => {common,database}
